EchoSrv: Guard backend initialisation with try/catch
